backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg[binary]>=3.1.18
python-decouple==3.8
Pillow>=10.4.0
django-filter==24.2
gunicorn==22.0.0
whitenoise==6.6.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
pymongo==4.8.0
daphne==4.1.2
pusher==3.3.3
